Initial commit
At present, the consistency of `typeintersect` and `issubtype` is
tested. This is done by randomly constructing a type `x`, and then
randomly widening/generalizing it to obtain two (probably different)
types `a` and `b` such that

```
x <: a && x <: b
```

It is then verified that

```
typeintersect(a, b) == typeintersect(b, a)    &&
x <: typeintersect(a, b)                      &&
typeintersect(a, b) <: a                      &&
typeintersect(a, b) <: b
```

The constructed types are made of nested `Tuple`s, `Val`s, `Int64`s,
`Float64`s, and integer constants. The random widening is performed by
introducing `TypeVar`s and wrapping in a `UnionAll` (with bounds
`Union{} <: T <: Any`) at the outmost level. Thus, the types may be
deeply nested, but are conceptually simple as they do not contain
`Union`s, `Vararg`s, or non-trivial `TypeVar` bounds.